C stat support

+++ scheme/resource.rs
@@ -74,7 +74,7 @@ impl Resource for DirResource {
 
     fn stat(&self, stat: &mut Stat) -> Result<usize> {
         stat.st_mode = MODE_DIR;
-        stat.st_size = self.data.len() as u64;
+        stat.st_size = self.data.len() as u32;
         Ok(0)
     }
 
@@ -144,7 +144,7 @@ impl Resource for FileResource {
 
     fn stat(&self, stat: &mut Stat) -> Result<usize> {
         stat.st_mode = MODE_FILE;
-        stat.st_size = self.size;
+        stat.st_size = self.size as u32;
         Ok(0)
     }
